Migrate types to carbon/index.d.ts
