+++ packages/carbon/src/commands/build.ts
@@ -13,11 +13,11 @@ import {
     getRoutes,
 } from "../actions";
 
-import { NodeActions } from "../types";
+import { Actions } from "../..";
 
 export default async ({ dir }: { dir: string }) => {
     loadConfig(dir);
-    const actions: NodeActions = {
+    const actions: Actions = {
         createNode,
         createNodeId,
         getNodes,
